Add files via upload
Crypto Aml Check Bot
A money laundering check(AML) is a type of identity check used to determine whether a person is involved in money laundering activities. Whether a person is involved in money laundering activities. It is a legal requirement for companies to conduct such checks on customers as part of anti-money laundering efforts.

About the program
Crypto Aml Checker Bot is a program that checks cryptocurrency transactions for compliance with Anti-Money Laundering (AML). It uses various algorithms to analyze transaction data, such as the size of the amount, the source and destination of funds, as well as information about the participants in the transaction, based on this data, the program will calculate the risk percentage. The main goal of Crypto Aml Checker Bot is to prevent the use of cryptocurrency involved in criminal fraud

  

Usage
After starting the program, select the type of wallet address of the wallet, to do this, press the key with the number of the desired type of address. Then enter the wallet address you want to verify or the transaction hash of where the recipient's address will be located. The program will begin to calculate the risk of the AML wallet address, after calculation the risk percentage from 0 to 100 will be displayed.